# api/api.yaml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: TorrPlay API
  description: A torrent streaming application with memory-managed piece storage.
  termsOfService: https://example.com/terms/
  contact:
    name: TorrPlay Team
    email: support@example.com
    url: https://github.com/torrplay
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
security:
  - basicAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/settings:
    get:
      tags:
        - v1
      summary: Get application settings.
      description: Retrieve current application configuration and settings.
      operationId: getSettings
      responses:
        "200":
          description: Settings retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    patch:
      tags:
        - v1
      summary: Update application settings.
      description: Update application configuration. Some settings may require a restart to take effect.
      operationId: updateSettings
      requestBody:
        description: New settings values.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "204":
          description: Settings updated successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
  responses:
...
    BadRequest:
      description: The request was invalid or cannot be served.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid info hash format."
            code: 400
            details: "Info hash must be 40 hex characters."
    Unauthorized:
      description: Authentication failed or is missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized."
            code: 401
            details: "Authentication token is missing or invalid."
...
    InternalServerError:
      description: An internal server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error."
            code: 500
            details: "Failed to allocate memory for piece storage."
...
  schemas:
    Auth:
      type: object
      description: Authentication settings.
      properties:
        enabled:
          type: boolean
          description: Enable or disable authentication.
          example: false
        type:
          type: string
          description: Authentication type.
          enum:
            - "basic"
            - "bearer"
          example: "basic"
        username:
          type: string
          description: Username for authentication.
          minimum: 4
          maximum: 64
          example: "admin"
        password:
          type: string
          description: Password for authentication.
          minimum: 4
          maximum: 128
          example: "password"
...
    Error:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
          description: Human-readable error message.
          example: "Torrent not found"
        code:
          type: integer
          description: HTTP status code.
          minimum: 400
          maximum: 599
          example: 404
...
    Settings:
      type: object
      description: Application configuration settings.
      properties:
        auth:
          $ref: "#/components/schemas/Auth"
        disable_ipv6:
          type: boolean
          description: Disable IPv6 for torrent client.
          example: false
        enable_dlna:
          type: boolean
          description: Enable DLNA media server functionality.
          example: false
        enable_downloader:
          type: boolean
          description: Enable background downloading for torrents with file storage.
          example: false
        file_storage_path:
          type: string
          description: Path to the file storage.
          example: "/torrplay"
        friendly_name:
          type: string
          description: DLNA friendly name displayed to clients.
          minLength: 3
          maxLength: 15
          example: "TorrPlay"
        http_server_port:
          type: integer
          description: TCP port for HTTP server.
          minimum: 1
          maximum: 65535
          example: 8090
        log_format:
          type: string
          description: Log format.
          enum:
            - "json"
            - "text"
          example: "text"
        log_level:
          type: string
          x-go-type: slog.Level
          x-go-type-import:
            path: log/slog
          description: Log level.
          example: "INFO"
        log_store_size:
          type: integer
          description: Number of log entries to store in memory.
          minimum: 0
          maximum: 1000
          example: 100
        max_memory:
          type: integer
          format: int64
          description: Maximum memory allocation for piece storage in bytes.
          minimum: 33554432
          maximum: 2147483648
          example: 67108864
        readahead_percentage:
          type: integer
          description: Readahead percentage value (50-100).
          minimum: 50
          maximum: 100
          example: 90
...